+++ home/bin/executable_mpv.sh
@@ -1,0 +1,9 @@
+#!/usr/bin/env sh
+
+set -o errexit
+
+mpv --no-config \
+    --geometry="50%x50%" \
+    --auto-window-resize=no \
+    --profile=fast \
+    --quiet $@