+++ home/scripts/gentoo.sh
@@ -5,73 +5,125 @@ set -o errexit -o nounset
 . $HOME/.local/share/chezmoi/home/scripts/lib.sh
 
 ins="emerge -av --changed-use"
-pkgs=""
-USE_DIR="/etc/portage/package.use"
+pkgs="app-crypt/gnupg pass stow dev-lang/ruby app-arch/tar app-arch/unzip app-arch/xz-utils net-misc/curl"
+user_groups=""
+services=""
+
 AUTH=$(search_auth)
 
-euse_pkg() {
-  if ! grep -q "$2" "$USE_DIR"/"${1#*/}" 2>/dev/null ; then
-    "$AUTH" euse -p "$1" -E "$2"
-  fi
+USE_DIR="/etc/portage/package.use"
+
+# Required portage dirs
+[ -d "$USE_DIR" ] || "$AUTH" mkdir -p "$USE_DIR"
+[ -d /etc/portage/package.accept_keyworks ] || "$AUTH" mkdir -p /etc/portage/package.accept_keywords
+
+# Make sure to install gentoolkit
+if ! hash euse 2>/dev/null; then
+    "$AUTH" $ins gentoolkit
+fi
+
+euse_enable() {
+  if "$AUTH" euse -E "$1" ; then echo "$1 enabled"; fi
 }
 
-euse_global() {
-  if ! grep -q "$1" /etc/portage/make.conf ; then
-    "$AUTH" euse -E "$1"
-  fi
+euse_disable() {
+  if "$AUTH" euse -D "$1" ; then echo "$1 disabled"; fi
 }
 
-install_deps() {
-  # env
-  if ! hash euse 2>/dev/null; then "$AUTH" $ins gentoolkit ; fi
-  euse_global jpeg
-
-  [ -d "$USE_DIR" ] || "$AUTH" mkdir -p "$USE_DIR"
-  [ -d /etc/portage/package.accept_keyworks ] || "$AUTH" mkdir -p /etc/portage/package.accept_keywords
-
-  "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.accept_keywords/* /etc/portage/package.accept_keywords/
-  "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/dotfiles "$USE_DIR/"
-
-  if has_systemd ; then
-    "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/lightdm-systemd "$USE_DIR/"
-    pkgs="brave-bin"
-  else
-    "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/lightdm-elogind "$USE_DIR/"
-    pkgs="librewolf"
-  fi
-
-  pkgs="$pkgs app-crypt/gnupg pass zsh awesome media-sound/mpd ncmpcpp xinit xorg-server xst
-    feh picom maim vifm mpv zathura zathura-pdf-mupdf
-    neomutt cava ueberzug weechat net-misc/yt-dlp
-    papirus-icon-theme media-sound/mpc inotify-tools light stow
-    arc-theme ffmpegthumbnailer tmux net-mail/isync xss-lock
-    app-misc/jq x11-misc/betterlockscreen app-shells/starship
-    dev-lang/ruby app-arch/tar app-arch/unzip app-arch/xz-utils"
+make_service() {
+    if has_systemd; then
+        "$AUTH" systemctl enable "$1"
+    else
+        "$AUTH" rc-update "$1" boot
+    fi
 }
 
-install_pulse() {
-  "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/pulseaudio "$USE_DIR/"
+add_awesome() {
+    euse_enable X
+    euse_disable wayland
 
-  pkgs="$pkgs pulseaudio"
+    "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/dotfiles "$USE_DIR/wm"
+
+    if has_systemd ; then
+        "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/lightdm-systemd "$USE_DIR/login"
+    else
+        "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/lightdm-elogind "$USE_DIR/login"
+    fi
+
+    # AppImage need fuse:0
+    pkgs="$pkgs awesome media-sound/mpd ncmpcpp xinit
+    xorg-server xst feh picom maim mpv zathura
+    zathura-pdf-mupdf neomutt cava weechat
+    papirus-icon-theme media-sound/mpc sys-fs/fuse:0
+    inotify-tools light net-mail/isync
+    xss-lock app-misc/jq x11-misc/betterlockscreen"
 }
 
-install_alsa() {
-  "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/alsa "$USE_DIR/"
+add_pulse() {
+    euse_enable pulseaudio
+    euse_disable alsa
+    euse_disable alsa-plugin
 
-  pkgs="$pkgs alsa-utils tap-plugins swh-plugins libsamplerate cmt-plugins ladspa-bs2b alsa-plugins"
+    "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/pulseaudio "$USE_DIR/sound"
+
+    pkgs="$pkgs pulseaudio nnn gnome-extra/nemo"
 }
 
-install_emacs() {
+add_alsa() {
+    euse_enable alsa
+    euse_disable pulseaudio
+
+    pkgs="$pkgs alsa-utils tap-plugins swh-plugins
+    libsamplerate cmt-plugins ladspa-bs2b alsa-plugins
+    nnn xfce-base/thunar xfce-base/tumbler"
+
+    if ! has_systemd; then
+        services="$services alsasound"
+    fi
+}
+
+add_swayfx() {
+    euse_enable wayland
+    euse_disable X
+
+    "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/swayfx "$USE_DIR/wm"
+
+    pkgs="$pkgs light papirus-icon-theme inotify-tools
+    swaybg imagemagick imv app-misc/jq playerctl
+    wl-clipboard gui-apps/grim gui-apps/wmenu
+    media-sound/mpc zathura wezterm
+    net-wireless/iwd gui-apps/eww gui-wm/swayfx
+    mpv-mpris mpd-mpris acct-group/seat seatd"
+
+    user_groups="$user_groups video seat"
+    services="$services seatd"
+}
+
+add_brave() {
+    pkgs="$pkgs brave-bin"
+}
+
+add_librewolf() {
+    pkgs="$pkgs librewolf-bin"
+}
+
+add_emacs() {
   "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/emacs "$USE_DIR/"
 
   # discount = markdown
   pkgs="$pkgs ripgrep discount emacs"
 }
 
-install_vim() {
-  "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.use/vim "$USE_DIR/"
+add_neovim() {
+    pkgs="$pkgs neovim sys-apps/fd ripgrep fzf tmux"
+}
 
-  pkgs="$pkgs vim"
+add_vim() {
+  pkgs="$pkgs vim tmux"
+}
+
+add_zsh() {
+    pkgs="$pkgs zsh app-shells/starship net-misc/yt-dlp"
 }
 
 install_extra_deps() {
@@ -80,48 +132,62 @@ install_extra_deps() {
 
 usage() {
   printf "\nUsage:\n"
-  echo " --deps         Install dependencies"
   echo " --sound-pulse  Install deps for PulseAudio"
   echo " --sound-alsa   Install deps for ALSA"
   echo " --extra-deps   Install other dependencies"
-  echo " --vim          Install deps for vim"
-  echo " --emacs        Install deps for emacs"
+  echo " --emacs        Install deps for DoomEmacs"
+  echo " --neovim       Install deps for NeoVim"
+  echo " --vim          Install deps for Vim"
+  echo " --awesome      Install deps for Awesome"
+  echo " --swayfx       Install deps for Swayfx"
+  echo " --brave        Install deps for Brave"
+  echo " --librewolf    Install deps for LibreWolf"
+  echo " --zsh          Install deps for Zsh"
 }
 
 ## CLI options
-DEPS=false
-PULSE=false
-ALSA=false
 EXTRA=false
-VIM=false
-EMACS=false
 
 if [ "$#" -eq 0 ] ; then usage ; exit 1 ; fi
 
 while [ "$#" -gt 0 ] ; do
-  case "$1" in
-    --deps) DEPS=true ;;
-    --sound-pulse) PULSE=true ;;
-    --sound-alsa) ALSA=true ;;
-    --extra-deps) EXTRA=true ;;
-    --vim) VIM=true ;;
-    --emacs) EMACS=true ;;
-    *) usage ; exit 1 ;;
-  esac
-  shift
+    case "$1" in
+        --sound-alsa) add_alsa ;;
+        --sound-pulse) add_pulse ;;
+        --extra-deps) EXTRA=true ;;
+        --awesome) add_awesome ;;
+        --swayfx) add_swayfx ;;
+        --emacs) add_emacs ;;
+        --neovim) add_neovim ;;
+        --vim) add_vim ;;
+        --brave) add_brave ;;
+        --librewolf) add_librewolf ;;
+        --zsh) add_zsh ;;
+        *) usage ; exit 1 ;;
+    esac
+    shift
 done
 
 main() {
-  "$DEPS" && install_deps
-  "$PULSE" && install_pulse
-  "$ALSA" && install_alsa
-  "$VIM" && install_vim
-  "$EMACS" && install_emacs
-  "$EXTRA" && install_extra_deps
+    "$AUTH" cp ~/.local/share/chezmoi/home/scripts/gentoo/package.accept_keywords/* /etc/portage/package.accept_keywords/
 
-  "$AUTH" $ins $pkgs
+    "$AUTH" $ins $pkgs
 
-  exit 0
+    "$EXTRA" && install_extra_deps
+
+    for group in $user_groups ; do
+        if [ "$group" != "" ] ; then
+            "$AUTH" usermod -aG "$group" "$USER"
+        fi
+    done
+
+    for service in $services ; do
+        if [ "$service" != "" ] ; then
+            make_service $service
+        fi
+    done
+
+    exit 0
 }
 
 main "$@"