+++ home/scripts/void.sh
@@ -3,8 +3,11 @@
 set -o errexit -o nounset
 
 . $HOME/.local/share/chezmoi/home/scripts/lib.sh
-ins="xbps-install -y"
-pkgs=""
+
+ins="xbps-install -Sy"
+pkgs="gnupg2 pass curl base-devel wget ruby unzip tar xz stow"
+services=""
+user_groups=""
 
 AUTH=$(search_auth)
 
@@ -12,30 +15,65 @@ build() {
   :
 }
 
-# Remove nerd-fonts-ttf (package size ~= 7G)
-install_deps() {
-  pkgs="xst cava ueberzug betterlockscreen gnupg2 pass xclip zsh
-    curl awesome mpd ncmpcpp xinit xorg-apps xorg-minimal xorg-input-drivers
-    base-devel wget feh picom maim vifm mpv zathura zathura-pdf-mupdf isync
-    neomutt ImageMagick weechat youtube-dl papirus-icon-theme mpc
-    inotify-tools light stow arc-theme
-    ffmpegthumbnailer tmux firefox xss-lock jq xrdb gcc starship ruby unzip tar xz"
+add_awesome() {
+    pkgs="xst cava betterlockscreen xclip awesome
+    mpd ncmpcpp xinit xorg-apps xorg-minimal
+    xorg-input-drivers feh picom maim mpv
+    zathura zathura-pdf-mupdf isync neomutt
+    ImageMagick weechat youtube-dl papirus-icon-theme
+    mpc inotify-tools light tmux xss-lock jq xrdb gcc"
+
+    services="$services acpid"
 }
 
-install_pulse() {
-  pkgs="$pkgs pulseaudio pulseaudio-equalizer-ladspa"
+add_pulse() {
+  pkgs="$pkgs pulseaudio pulseaudio-equalizer-ladspa nnn nemo"
+
+  user_groups="$user_groups dbus"
 }
 
-install_alsa() {
-  pkgs="$pkgs alsa-utils alsa-plugins ladspa-bs2b swh-plugins alsa-plugins-samplerate"
+add_alsa() {
+  pkgs="$pkgs alsa-utils alsa-plugins ladspa-bs2b swh-plugins alsa-plugins-samplerate nnn Thunar"
+
+  user_groups="$user_groups audio"
+
+  services="$services alsa"
 }
 
-install_emacs() {
-  pkgs="$pkgs ripgrep emacs-gtk3"
+add_swayfx() {
+    pkgs="$pkgs swayfx imv light jq wl-clipboard
+    papirus-icon-theme inotify-tools mpd mpc wezterm
+    playerctl mpv-mpris mpDris2 eww swaybg grim wmenu
+    iwd seatd turnstile"
+
+    user_groups="$user_groups _seatd"
+
+    services="$services seatd turnstiled"
 }
 
-install_vim() {
-  pkgs="$pkgs vim vim-x11"
+add_brave() {
+    :
+}
+
+# added with appimage if web = "librewolf" is set
+add_librewolf() {
+    :
+}
+
+add_emacs() {
+    pkgs="$pkgs ripgrep emacs-gtk3"
+}
+
+add_neovim() {
+    pkgs="$pkgs neovim fd fzf tmux gcc"
+}
+
+add_vim() {
+    pkgs="$pkgs vim vim-x11"
+}
+
+add_zsh() {
+    pkgs="$pkgs starship zsh"
 }
 
 install_extra_deps() {
@@ -47,58 +85,61 @@ make_service() {
   [ -s /var/service/"$1" ] || "$AUTH" ln -s /etc/sv/"$1" /var/service/"$1"
 }
 
-services() {
-  if "$ALSA" ; then make_service alsa ; fi
-  make_service acpid
-}
-
 usage() {
   printf "\nUsage:\n"
-  echo " --deps         Install dependencies"
   echo " --sound-pulse  Install deps for PulseAudio"
   echo " --sound-alsa   Install deps for ALSA"
   echo " --extra-deps   Install other dependencies"
-  echo " --vim          Install deps for vim"
-  echo " --emacs        Install deps for emacs"
+  echo " --emacs        Install deps for DoomEmacs"
+  echo " --neovim       Install deps for NeoVim"
+  echo " --vim          Install deps for Vim"
+  echo " --awesome      Install deps for Awesome"
+  echo " --swayfx       Install deps for Swayfx"
+  echo " --brave        Install deps for Brave"
+  echo " --librewolf    Install deps for LibreWolf"
+  echo " --zsh          Install deps for Zsh"
 }
 
 ## CLI options
-DEPS=false
-PULSE=false
-ALSA=false
 EXTRA=false
-VIM=false
-EMACS=false
 
 if [ "$#" -eq 0 ] ; then usage ; exit 1 ; fi
 
 while [ "$#" -gt 0 ] ; do
-  case "$1" in
-    --deps) DEPS=true ;;
-    --sound-pulse) PULSE=true ;;
-    --sound-alsa) ALSA=true ;;
-    --extra-deps) EXTRA=true ;;
-    --vim) VIM=true ;;
-    --emacs) EMACS=true ;;
-    *) usage; exit 1 ;;
-  esac
-  shift
+    case "$1" in
+        --sound-alsa) add_alsa ;;
+        --sound-pulse) add_pulse ;;
+        --extra-deps) EXTRA=true ;;
+        --awesome) add_awesome ;;
+        --swayfx) add_swayfx ;;
+        --emacs) add_emacs ;;
+        --neovim) add_neovim ;;
+        --vim) add_vim ;;
+        --brave) add_brave ;;
+        --librewolf) add_librewolf ;;
+        --zsh) add_zsh ;;
+        *) usage; exit 1 ;;
+    esac
+    shift
 done
 
 main() {
-  "$DEPS" && install_deps
-  "$PULSE" && install_pulse
-  "$ALSA" && install_alsa
-  "$VIM" && install_vim
-  "$EMACS" && install_emacs
+    "$AUTH" $ins $pkgs
+    "$EXTRA" && install_extra_deps
 
-  "$AUTH" $ins $pkgs
+    for group in $user_groups ; do
+        if [ "$group" != "" ] ; then
+            "$AUTH" usermod -aG "$group" "$USER"
+        fi
+    done
 
-  "$EXTRA" && install_extra_deps
+    for service in $services ; do
+        if [ "$service" != "" ] ; then
+            make_service $service
+        fi
+    done
 
-  services
-
-  exit 0
+    exit 0
 }
 
 main "$@"